+++ NEWS.md
@@ -3,6 +3,11 @@
 The noteworthy changes for each Clearance version are included here. For a
 complete changelog, see the git history for each version via the version links.
 
+### Fixed
+
+- Delete cookie correctly when a callable object is set as the custom domain
+  setting.
+
 ## [2.2.0] - July 9, 2020
 
 ### Added